meet-guide-components/mispronunciation-detection-system/requirements.txt
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.20.0
ffmpeg-python>=0.2.0

# API & Environment
python-dotenv>=1.0.0
deepgram-sdk>=3.0.0
requests>=2.28.0

# Pronunciation Detection Dependencies
transformers>=4.30.0
torchaudio>=2.0.0
nltk>=3.8.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
textgrid>=1.5.0
matplotlib>=3.7.0

# Audio-Aware Correction Dependencies
sentence-transformers>=2.2.0
sentencepiece>=0.1.99
protobuf>=3.20.0
